Change model filename to signalfx_metrics
This is necessary to prevent a conflict with Jaeger which also uses the
file "model.proto", which conflicts with the old name we used.  See
https://developers.google.com/protocol-buffers/docs/reference/go/faq#fix-namespace-conflict
for more details on why this is bad.

Also fix issue with path to protobuf includes when using modules.

This is only backwards-imcompatible under specific circumstances like
storing our protobuf models in an Any field and accessing it at a later
time with an upgraded version of this library.  That is not an expected
use of these models so the risk seems acceptable.